requirements.txt
flask-restx==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
flask-marshmallow==1.3.0
flask-cors==5.0.0
marshmallow==3.25.1
PyMySQL==1.1.1
Python==3.12.7
PyJWT==2.8.0
python-dotenv==0.21.0
Werkzeug==3.0.3
pipreqs==0.5.0
